Move glassfish 4 compatibility layer to instrumentation module
