+++ settings.gradle
@@ -40,6 +40,7 @@ include ':dd-java-agent:instrumentation:elasticsearch:transport-2'
 include ':dd-java-agent:instrumentation:elasticsearch:transport-5'
 include ':dd-java-agent:instrumentation:elasticsearch:transport-5.3'
 include ':dd-java-agent:instrumentation:elasticsearch:transport-6'
+include ':dd-java-agent:instrumentation:glassfish-4.1'
 include ':dd-java-agent:instrumentation:grpc-1.5'
 include ':dd-java-agent:instrumentation:hibernate'
 include ':dd-java-agent:instrumentation:hibernate:core-3.3'
